[ADD] 자동 로그인 시 메인으로 이동
